Finished translation - first draft - Passing simple manifold tests - Passing Chicago dataset tests - Currently using scipy.optimize.minimize as nls solver - Potential bug identified, res$summary varies from R implementation
