+++ fastEDM/__init__.py
@@ -1,2 +1,3 @@
 from _fastEDM import run_command
 from .edm import edm, create_manifold
+from .easy_edm import easy_edm
