+++ pyproject.toml
@@ -2,7 +2,8 @@
 requires = [
     "setuptools>=42",
     "pybind11>=2.10.0",
-    "numpy>=1.17.0"
+    "numpy>=1.17.0",
+    "scipy>=1.9.2"
 ]
 build-backend = "setuptools.build_meta"
